Filter out noscript tags from SharePoint pages
b/116484667

SharePoint pages include noscript tags with warning text message
"You may be trying to access this site from a secured browser on the server."
This message appears in generated snippets for search results. So we are
taking it out.

Also page contains navigation items which are inside <div> with CSS
class "noIndex". We are filtering out navigation sections so values from
navigation do not match to each SharePoint page.

Change-Id: I97d5701ef8ad1a0f439bf4937e47f7dd4ec46556